refactor: eco 패키지 분리 및 불필요한 item 필드 제거 (#2)
