--- src/main/java/com/chocobi/leafy/place/dto/EcoItem.java
@@ -1,15 +1,0 @@
-package com.chocobi.leafy.place.dto;
-
-import lombok.Data;
-
-@Data
-public class EcoItem {
-    private String addr;
-    private String areacode;
-    private String sigungucode;
-    private String mainimage;
-    private String title;
-    private String summary;
-    private String tel;
-    private String cpyrhtDivCd;
-}
